ai/requirements.txt
fastapi==0.115.5
pydantic==2.10.3
uvicorn[standard]==0.32.1
torch>=2.1.0
torchvision>=0.16.0
onnxruntime>=1.17.0
pillow>=10.0.0
numpy>=1.26.0
python-multipart>=0.0.12
requests>=2.31.0
transformers>=4.40.0
accelerate>=0.20.0
sentencepiece>=0.2.0
llama-cpp-python>=0.2.90
huggingface-hub>=0.20.0
